+++ apps/ui-dev/pages/sapa/feature-jsx.jsx
@@ -9,8 +9,8 @@ import vnodeJsText from "./codes/vnode.js.txt?raw";
 function Main() {
   return (
     <Layout>
-      <h1>Template - vnode</h1>
-      <p>Sapa 는 기본적으로 virtual dom 을 씁니다.</p>
+      <h1>Jsx</h1>
+      <p>Sapa 는 기본 템플릿으로 jsx를 씁니다.</p>
       <p>React 와 같은 방식의 Jsx 문법을 통해서 사용합니다.</p>
 
       <p>