+++ packages/ui/index.d.ts
@@ -119,6 +119,26 @@ declare module "@elf-framework/ui" {
     wrap?: string;
   }
 
+  export type InputPropsType = {
+    type?: string;
+    value?: string;
+    placeholder?: string;
+    selected?: boolean;
+    checked?: boolean;
+    disabled?: boolean;
+    readOnly?: boolean;
+    required?: boolean;
+    maxLength?: number;
+    minLength?: number;
+    max?: number;
+    min?: number;
+    step?: number;
+    pattern?: string;
+    size?: number;
+    rows?: number;
+    cols?: number;
+  };
+
   export type DomEventType = {
     onFocus?: (event: KeyboardEvent) => void;
     onBlur?: (event: KeyboardEvent) => void;
@@ -678,10 +698,12 @@ declare module "@elf-framework/ui" {
 
   export interface ViewProps {
     as?: string;
+    id?: string;
+    class?: string;
     style?: CommonStyle;
   }
   export class View extends UIElement {
-    props: ViewProps & CommonStyle & DomEventType;
+    props: ViewProps & CommonStyle & DomEventType & InputPropsType;
   }
 
   export type InputEditorStyle = {