+++ packages/ui/src/components/view/index.jsx
@@ -1,16 +1,25 @@
-import { UIElement } from "@elf-framework/sapa";
+import { classnames, UIElement } from "@elf-framework/sapa";
 
 import { propertyMap } from "../../utils/propertyMap";
 import { convertPropertyToStyleKey } from "../../utils/styleKeys";
 
 export class View extends UIElement {
   template() {
-    const { as = "div", style = {}, content, ...extraStyle } = this.props;
+    const {
+      as = "div",
+      id,
+      class: className = "",
+      style = {},
+      content,
+      ...extraStyle
+    } = this.props;
 
     const { style: styleProperties, noneStyle } =
       convertPropertyToStyleKey(extraStyle);
 
     const styleObject = {
+      class: classnames(className),
+      id,
       style: propertyMap({ ...style, ...styleProperties }, {}),
       ...noneStyle,
     };